requirements.txt
bioimageio.spec @ git+https://github.com/bioimage-io/spec-bioimage-io@cfacdc292784e0ecc7f5dbcdce0e4fc2b0ca3cad  # TODO: chenage to released version
loguru==0.7.2
minio==7.2.3
packaging==23.2
ruyaml==0.91.0
spdx-license-list==3.22
typer==0.9.0
